requirements.txt
# Django & Core
Django==1.11.9
pymysql==0.8.0
django-material==1.2.2
urllib3==1.22
django-crispy-forms==1.7.0
django-watson==1.5.0
django_compressor==2.2
django-cors-headers==2.1.0
django-stronghold==0.3.0

social-auth-app-django==2.1.0
# OLD: python-social-auth==0.2.21

jsonfield==2.0.2

# Debuging & Logging
colorlog==3.1.0
django-debug-toolbar==1.9.1
django-extensions==1.9.9
Werkzeug==0.14.1

# Documentation
sphinx
sphinx_rtd_theme
django-docs==0.2.1

# API
djangorestframework==3.7.7
pygments==2.2.0

# Testing and tools
Faker

# Server
gunicorn
gevent
